CASM-4729: fix
